changed package + refactoring
